+++ tests/data/torch_test/src/lib.rs
@@ -1,0 +1,8 @@
+use tch::Tensor;
+
+#[test]
+fn test_dummy() {
+    let t = Tensor::of_slice(&[3, 1, 4, 1, 5]);
+    let t = t * 2;
+    t.print();
+}