+++ travis-install.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-curl -sL https://github.com/xd009642/tarpaulin/releases/download/0.18.5/cargo-tarpaulin-0.18.5-travis.tar.gz | tar xvz -C $HOME/.cargo/bin
+curl -sL https://github.com/xd009642/tarpaulin/releases/download/0.19.0/cargo-tarpaulin-0.19.0-travis.tar.gz | tar xvz -C $HOME/.cargo/bin