+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 From 2019 onwards, all notable changes to tarpaulin will be documented in this
 file.
 
+## [0.19.0] 2021-12-27
+### Added
+- Check build script output from cargo build and use it to set `LD_LIBRARY_PATH` to match cargo test behaviour
+- `--implicit-test-flags` argument so that `--test-threads` isn't passed into the test binary
+
+### Changed
+- Parse RUSTFLAGS and RUSTDOCFLAGS to remove duplicate entries #891
+- Explicitly pass `--test-threads` to test binary to counteract cpu affinity being set to 1 CPU
+
+### Removed
+
 ## [0.18.5] 2021-11-05
 ### Added